projects/snowfalke-sync/requirements.txt
Flask==2.3.3
pandas==2.1.1
snowflake-connector-python==3.4.0
PyYAML==6.0.1
Werkzeug==2.3.7